refactor(core): extract node-credential-requirement helpers into a focused module
The credential-introspection logic (computeNodesRequiringCredentialSetup
+ three private helpers) was ~70 lines of pure node-type analysis sitting
in the middle of DesktopAssistantService. It only depends on NodeTypes
+ NodeHelpers, not on any other service state.

Move it to node-credential-requirements.ts as module-level pure functions
that take NodeTypes as a parameter. The service drops the four private
methods and calls one function from getTasks:

  nodesRequiringCredentialSetup: computeNodesRequiringCredentialSetup(
    workflow.nodes,
    this.nodeTypes,
  ),

Behavior preserved; existing 60 tests still pass.
