+++ packages/cli/src/modules/instance-ai/desktop-assistant/desktop-assistant.service.ts
@@ -17,14 +17,7 @@ import { Service } from '@n8n/di';
 import type { StoredEvent } from '@n8n/instance-ai';
 // eslint-disable-next-line n8n-local-rules/misplaced-n8n-typeorm-import
 import { In } from '@n8n/typeorm';
-import type {
-	ExecutionStatus,
-	INode,
-	INodeCredentialDescription,
-	INodeParameters,
-	INodeTypeDescription,
-} from 'n8n-workflow';
-import { NodeHelpers } from 'n8n-workflow';
+import type { ExecutionStatus } from 'n8n-workflow';
 import { randomUUID } from 'node:crypto';
 
 import { CredentialsFinderService } from '@/credentials/credentials-finder.service';
@@ -41,6 +34,7 @@ import {
 } from './constants';
 import { classifyWorkflowsForDesktopAssistant } from './desktop-assistant.classifier';
 import type { ClassifierInput } from './desktop-assistant.classifier';
+import { computeNodesRequiringCredentialSetup } from './node-credential-requirements';
 
 import { InProcessEventBus } from '../event-bus/in-process-event-bus';
 import { InstanceAiMemoryService } from '../instance-ai-memory.service';
@@ -162,99 +156,14 @@ export class DesktopAssistantService {
 					: undefined,
 				emojiIcon: metaIcon ?? nameEmoji,
 				accessibleCredentialIds,
-				nodesRequiringCredentialSetup: this.computeNodesRequiringCredentialSetup(workflow.nodes),
+				nodesRequiringCredentialSetup: computeNodesRequiringCredentialSetup(
+					workflow.nodes,
+					this.nodeTypes,
+				),
 			};
 		});
 
 		return classifyWorkflowsForDesktopAssistant(inputs);
-	}
-
-	/**
-	 * Identify nodes whose type declares required credentials but whose workflow
-	 * JSON has no populated credential slot. Reuses the same primitives the
-	 * workflow-builder's `setup-workflow.service.ts` analysis is built on —
-	 * `INodeTypeDescription.credentials` filtered by `NodeHelpers.displayParameter`
-	 * against the current parameters — but stays synchronous and avoids the
-	 * heavier `buildSetupRequests` pipeline (parameter issues, credential cache,
-	 * testability checks). For the BFF classifier we just need the "missing
-	 * setup" signal.
-	 *
-	 * The existing slot-based detection (empty `slot.id` or inaccessible
-	 * credential) continues to live in the classifier itself; this complements
-	 * it by catching the case where the node has no `credentials` field at all
-	 * (a workflow imported / built without ever picking a credential).
-	 */
-	private computeNodesRequiringCredentialSetup(nodes: INode[]): Set<string> {
-		const result = new Set<string>();
-		for (const node of nodes) {
-			if (node.disabled) continue;
-			// If any credential slot already exists, the classifier's slot-based
-			// check handles whether it's populated correctly.
-			if (node.credentials && Object.keys(node.credentials).length > 0) continue;
-
-			const description = this.tryGetNodeDescription(node);
-			if (!description) continue;
-			if (!Array.isArray(description.credentials) || description.credentials.length === 0) continue;
-
-			// Resolve defaults from the description before evaluating displayOptions.
-			// Many nodes (e.g. Gmail Trigger) gate their credentials on parameters
-			// whose values come from the description's default rather than being
-			// stored on the workflow JSON. Mirrors `resolveDisplayedDefaults` in
-			// the adapter service's getNodeCredentialTypes.
-			const resolvedParameters = this.resolveParametersWithDefaults(node, description);
-			const nodeWithDefaults: INode = { ...node, parameters: resolvedParameters };
-
-			const hasRequired = description.credentials.some((credentialDesc) =>
-				this.nodeRequiresCredential(credentialDesc, nodeWithDefaults, description),
-			);
-			if (hasRequired) result.add(node.name);
-		}
-		return result;
-	}
-
-	private tryGetNodeDescription(node: INode): INodeTypeDescription | undefined {
-		try {
-			return this.nodeTypes.getByNameAndVersion(node.type, node.typeVersion).description;
-		} catch {
-			return undefined;
-		}
-	}
-
-	private resolveParametersWithDefaults(
-		node: INode,
-		description: INodeTypeDescription,
-	): INodeParameters {
-		const parameters: INodeParameters = node.parameters ?? {};
-		try {
-			const resolved = NodeHelpers.getNodeParameters(
-				description.properties,
-				parameters,
-				true,
-				false,
-				node,
-				description,
-			);
-			return resolved ?? parameters;
-		} catch {
-			return parameters;
-		}
-	}
-
-	private nodeRequiresCredential(
-		credentialDesc: INodeCredentialDescription,
-		node: INode,
-		description: INodeTypeDescription,
-	): boolean {
-		if (credentialDesc.required === false) return false;
-		if (!credentialDesc.displayOptions) return true;
-		try {
-			return NodeHelpers.displayParameter(node.parameters, credentialDesc, node, description);
-		} catch {
-			// If displayOptions evaluation throws (malformed parameters etc.),
-			// be conservative and treat the credential as required so the user
-			// is nudged to inspect it.
-			return true;
-		}
 	}
 
 	private async fetchLastExecutionByWorkflowId(
